@@ -11,43 +11,58 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.0.2",
-    "@angular/common": "^6.0.2",
-    "@angular/compiler": "^6.0.2",
-    "@angular/core": "^6.0.2",
-    "@angular/forms": "^6.0.2",
-    "@angular/http": "^6.0.2",
-    "@angular/platform-browser": "^6.0.2",
-    "@angular/platform-browser-dynamic": "^6.0.2",
-    "@angular/router": "^6.0.2",
-    "@ng-bootstrap/ng-bootstrap": "^2.0.0",
-    "@types/dhtmlxscheduler": "^4.3.36",
-    "core-js": "^2.5.4",
+    "@angular/animations": "12.0.3",
+    "@angular/cdk": "^12.0.3",
+    "@angular/common": "12.0.3",
+    "@angular/compiler": "12.0.3",
+    "@angular/core": "^12.0.3",
+    "@angular/forms": "12.0.3",
+    "@angular/http": "^7.2.16",
+    "@angular/localize": "12.0.3",
+    "@angular/material": "^12.0.3",
+    "@angular/platform-browser": "^12.0.3",
+    "@angular/platform-browser-dynamic": "^12.0.3",
+    "@angular/router": "^12.0.3",
+    "@fortawesome/fontawesome-free": "^5.15.3",
+    "@ng-bootstrap/ng-bootstrap": "^9.1.2",
+    "@ng-select/ng-select": "^6.1.0",
+    "@types/chart.js": "^2.9.32",
+    "@types/dhtmlxscheduler": "^4.3.40",
+    "angular-alert-module": "^2.0.3",
+    "angular-bootstrap-md": "^11.1.0",
+    "angular-chart.js": "^1.1.1",
+    "animate.css": "^4.1.1",
+    "bootstrap": "^4.5.0",
+    "chart.js": "^2.5.0",
+    "chartjs-plugin-datalabels": "^1.0.0",
+    "core-js": "^3.14.0",
     "dhtmlx-scheduler": "^5.0.1",
+    "dhtmlx-suite": "^5.0.1",
+    "hammerjs": "^2.0.8",
     "ng-bootstrap": "^1.6.3",
-    "rxjs": "^6.0.0",
-    "rxjs-compat": "^6.2.0",
-    "zone.js": "^0.8.26"
+    "rxjs": "^6.6.7",
+    "rxjs-compat": "^6.6.7",
+    "zone.js": "^0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.3",
-    "@angular/cli": "~6.0.3",
-    "@angular/compiler-cli": "^6.0.2",
-    "@angular/language-service": "^6.0.2",
-    "@types/jasmine": "~2.8.6",
+    "@angular-devkit/build-angular": "12.0.3",
+    "@angular/cli": "12.0.3",
+    "@angular/compiler-cli": "12.0.3",
+    "@angular/language-service": "^12.0.3",
+    "@types/jasmine": "^3.7.7",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
+    "@types/node": "^8.9.5",
     "codelyzer": "~4.2.1",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
+    "karma": "^6.3.3",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~1.4.2",
     "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "~5.3.0",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "^2.7.2"
+    "ts-node": "^10.0.0",
+    "tslint": "^6.1.3",
+    "typescript": "^4.2.3"
   }
 }
