@@ -6,6 +6,8 @@ import { ViewChild } from '@angular/core';
 import { routes } from './app-routing.module';
 
 import 'rxjs/add/operator/filter';
+import { Patient } from './shared/patients/patient.model';
+import { PatientsService } from './shared/patients/patients.service';
 
 @Component({
   selector: 'app-root',
@@ -15,13 +17,15 @@ import 'rxjs/add/operator/filter';
 export class AppComponent implements OnInit {
   userClaims: any;
   title = 'mmm';
+  selectedPatientId: any;
+  patients: Patient[];
 
   public topNavLinks: Array<{
     path: string,
     name: string
   }> = [];
 
-  constructor(private router: Router, private userService: UserService) {
+  constructor(private router: Router, private userService: UserService, private patientsService: PatientsService) {
     this.userService.authChanged.subscribe((someData?: any) => {
       if (someData) {
         this.userService.getUserClaims().subscribe((data: any) => {
@@ -31,7 +35,15 @@ export class AppComponent implements OnInit {
     });
   }
 
+  selectedPatientChanged() {
+    if (this.selectedPatientId == null) {
+      this.selectedPatientId = 1;
+    }
+    this.router.navigate(['/patients/' + this.selectedPatientId]);
+  }
+
   ngOnInit() {
+    this.patientsService.getAllPatients().subscribe((data: any) => this.patients = data);
     this.userService.authChanged.subscribe((someData?: any) => {
       this.userClaims = someData;
     });
