@@ -7,14 +7,14 @@ import { Observable, throwError } from 'rxjs';
 })
 @Injectable()
 export class AuthGuard implements CanActivate {
-  constructor(private router: Router) {}
+  constructor(private router: Router) { }
   canActivate(
     next: ActivatedRouteSnapshot,
-    state: RouterStateSnapshot):  boolean {
-      if (localStorage.getItem('userToken') != null) {
+    state: RouterStateSnapshot): boolean {
+    if (localStorage.getItem('userToken') != null) {
       return true;
-      }
-      this.router.navigate(['/login']);
-      return false;
+    }
+    this.router.navigate(['/login']);
+    return false;
   }
 }
