@@ -5,6 +5,7 @@ import { Injectable, Inject } from '@angular/core';
 import { Router } from '@angular/router';
 import { of } from 'rxjs';
 import { tap, map } from 'rxjs/operators';
+import 'rxjs/add/operator/do';
 import { HttpClient } from '@angular/common/http';
 
 @Injectable()
@@ -13,25 +14,21 @@ export class AuthInterceptor implements HttpInterceptor {
     constructor(private router: Router) { }
 
     intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
-        if (req.headers.get('No-Auth') === 'True') {
-            return next.handle(req.clone());
-        }
-
         if (localStorage.getItem('userToken') != null) {
             const clonedreq = req.clone({
                 headers: req.headers.set('Authorization', 'Bearer ' + localStorage.getItem('userToken'))
             });
             return next.handle(clonedreq)
-            .pipe(
-                tap(succ => {  },
+                .do(
+                    succ => { },
                     err => {
                         if (err.status === 401) {
-                            this.router.navigateByUrl('/login');
+                            localStorage.removeItem('userToken');
                         }
-                    })
-            );
+                    }
+                );
         } else {
-            this.router.navigateByUrl('/login');
+            return next.handle(req.clone());
         }
     }
 }
