@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { Router, ActivatedRoute } from '@angular/router';
+import { Router, RouterModule, ActivatedRoute } from '@angular/router';
 import { UserService } from '../shared/user.service';
 import { DoctorsService } from '../shared/doctors/doctors.service';
 import { NgModule } from '@angular/core';
@@ -13,6 +13,11 @@ import { ImagesService } from '../shared/images/images.service';
 export class HomeComponent implements OnInit {
   userClaims: any;
   doctors: any;
+  // pager object
+  pager: any = {};
+
+  // paged items
+  pagedItems: any[];
 
   // tslint:disable-next-line:max-line-length
   constructor(private router: Router, private userService: UserService, private doctorsService: DoctorsService, private imagesService: ImagesService) {
