@@ -17,18 +17,31 @@ namespace HospitalReception.Controllers.Api
         private readonly IEntityBaseRepository<DisabilityGroup> _disabilityGroupsRepository;
         private readonly IEntityBaseRepository<InformationSource> _informationSourcesRepository;
         private readonly IEntityBaseRepository<HabitationMember> _habitationMembersRepository;
+        private readonly IEntityBaseRepository<LocalityType> _localityTypesRepository;
+        private readonly IEntityBaseRepository<Region> _regionsRepository;
+        private readonly IEntityBaseRepository<Organization> _organizationsRepository;
+        private readonly IEntityBaseRepository<EmploymentType> _employmentTypesRepository;
 
         public LookupsController(IEntityBaseRepository<Gender> genderRepository,
             IEntityBaseRepository<EducationType> educationTypesRepository,
             IEntityBaseRepository<DisabilityGroup> disabilityGroupsRepository,
             IEntityBaseRepository<InformationSource> informationSourcesRepository,
-            IEntityBaseRepository<HabitationMember> habitationMembersRepository)
+            IEntityBaseRepository<HabitationMember> habitationMembersRepository,
+            IEntityBaseRepository<LocalityType> localityTypesRepository,
+            IEntityBaseRepository<Region> regionsRepository,
+            IEntityBaseRepository<Organization> organizationsRepository,
+            IEntityBaseRepository<EmploymentType> employmentTypesRepository
+            )
         {
             _genderRepository = genderRepository;
             _educationTypesRepository = educationTypesRepository;
             _disabilityGroupsRepository = disabilityGroupsRepository;
             _informationSourcesRepository = informationSourcesRepository;
             _habitationMembersRepository = habitationMembersRepository;
+            _localityTypesRepository = localityTypesRepository;
+            _regionsRepository = regionsRepository;
+            _organizationsRepository = organizationsRepository;
+            _employmentTypesRepository = employmentTypesRepository;
         }
 
         [Route("genders")]
@@ -45,6 +58,14 @@ namespace HospitalReception.Controllers.Api
             var educationTypes = _educationTypesRepository.GetAll().ToList();
             var educationTypesViewModel = Mapper.Map<IEnumerable<EducationType>, IEnumerable<EducationTypeViewModel>>(educationTypes);
             return Ok(educationTypesViewModel);
+        }
+
+        [Route("employmentTypes")]
+        public IHttpActionResult GetEmploymentTypes()
+        {
+            var employmentTypes = _employmentTypesRepository.GetAll().ToList();
+            var employmentTypesViewModel = Mapper.Map<IEnumerable<EmploymentType>, IEnumerable<EmploymentTypeViewModel>>(employmentTypes);
+            return Ok(employmentTypesViewModel);
         }
 
         [Route("disabilityGroups")]
@@ -69,6 +90,30 @@ namespace HospitalReception.Controllers.Api
             var habitationMembers = _habitationMembersRepository.GetAll().ToList();
             var habitationMembersViewModel = Mapper.Map<IEnumerable<HabitationMember>, IEnumerable<HabitationMemberViewModel>>(habitationMembers);
             return Ok(habitationMembersViewModel);
+        }
+
+        [Route("policlinics")]
+        public IHttpActionResult GetPoliclinics()
+        {
+            var organizations = _organizationsRepository.GetAll().ToList();
+            var organizationsViewModel = Mapper.Map<IEnumerable<Organization>, IEnumerable<OrganizationViewModel>>(organizations);
+            return Ok(organizationsViewModel);
+        }
+
+        [Route("localityTypes")]
+        public IHttpActionResult GetLocalityTypes()
+        {
+            var localityTypes = _localityTypesRepository.GetAll().ToList();
+            var localityTypesViewModel = Mapper.Map<IEnumerable<LocalityType>, IEnumerable<LocalityTypeViewModel>>(localityTypes);
+            return Ok(localityTypesViewModel);
+        }
+
+        [Route("regions")]
+        public IHttpActionResult GetRegions()
+        {
+            var regions = _regionsRepository.GetAll().ToList();
+            var regionsViewModel = Mapper.Map<IEnumerable<Region>, IEnumerable<RegionViewModel>>(regions);
+            return Ok(regionsViewModel);
         }
     }
 }
